requirements.txt
apache-airflow==2.4.1
beautifulsoup4==4.11.1
itemadapter==0.7.0
psycopg2==2.9.4
requests==2.28.1
scrapy==2.6.3